+++ Code/m/src/m/serializing/UnitySerializer.java
@@ -27,6 +27,8 @@ import m.csharp.Default;
 import m.csharp.ExpressionStatement;
 import m.csharp.Field;
 import m.csharp.FloatLiteral;
+import m.csharp.Foreach;
+import m.csharp.Lambda;
 import m.csharp.MemberInitializer;
 import m.csharp.Method;
 import m.csharp.Modifier;
@@ -79,6 +81,7 @@ import m.library.Component;
 import m.library.SimpleType;
 import m.library.Type;
 import m.yaml.YamlFactory;
+import static m.serializing.CSharpHelper.*;
 
 public class UnitySerializer
 {
@@ -117,9 +120,6 @@ public class UnitySerializer
 				serialize(component, type);
 			}
 		}
-		serialize("Collisions", entityList);
-		serialize("CollisionEntries", entityList);
-		serialize("CollisionExits", entityList);
 		
 		for (var archetype : game.getArchetypes())
 		{
@@ -135,7 +135,7 @@ public class UnitySerializer
 		extras();
 		systemGroups();
 		inputSystem();
-		//Physics();
+		Physics();
 		timers();
 	}
 	
@@ -1252,7 +1252,7 @@ public class UnitySerializer
 			csExpression.setLeft(component);
 			csExpression.setRight(field);
 			component.setName(access.getComponent()+"_"+access.getEntity());
-			field.setName(field(access.getComponent()));
+			field.setName(fieldName(access.getComponent()));
 			if (MValidator.components.get(access.getComponent()) == entityList)
 			{
 				return component;
@@ -1379,7 +1379,7 @@ public class UnitySerializer
 		return result;
 	}
 	
-	private String field(String component)
+	private String fieldName(String component)
 	{
 		var engineComponents = Component.values();
 		for (var engineComponent : engineComponents)
@@ -1665,6 +1665,10 @@ public class UnitySerializer
 	
 	private void Physics()
 	{
+		serialize("Collisions", entityList);
+		serialize("CollisionEntries", entityList);
+		serialize("CollisionExits", entityList);
+		
 		var unit = unit();
 		unit.getUsings().add(namespaceUsing("UnityEngine"));
 		unit.getUsings().add(namespaceUsing("Unity.Physics"));
@@ -1691,7 +1695,7 @@ public class UnitySerializer
 		world.getAttributes().add(attribute("ReadOnly"));
 		
 		job.getMembers().add(world);
-		job.getMembers().add(field(new Modifier[] {PUBLIC}, "NativeList<CollisionData>", declarator("collsionData")));
+		job.getMembers().add(field(new Modifier[] {PUBLIC}, "NativeList<CollisionData>", declarator("collisionData")));
 		
 		var execute = method(new Modifier[] {PUBLIC}, "void", "Execute", new Parameter[] {parameter("CollisionEvent", "collisionEvent")});
 		job.getMembers().add(execute);
@@ -1705,377 +1709,50 @@ public class UnitySerializer
 		onCreate.getStatements().add(statement(assignment(variable("step"),access(variable("World"),function("GetOrCreateSystem", new String[] {"StepPhysicsWorld"})))));
 
 		var onUpdate = method(new Modifier[] {PROTECTED,OVERRIDE}, "JobHandle", "OnUpdate", new Parameter[] {parameter("JobHandle", "inputDependencies")});
+		
+		var clearLambda = lambda(new Parameter[] {parameter("DynamicBuffer<Collisions>", "buffer")});
+		clearLambda.getStatements().add(statement(access(variable("buffer"),function("Clear"))));
+		
+		var clearRun = access(access(access(variable("Entities"),function("ForEach",argument(clearLambda))),function("WithoutBurst")), function("Run"));
+		onUpdate.getStatements().add(statement(clearRun));
+		
+		var lambda = lambda(new Parameter[] {refParameter("ISimulation", "simulation"), refParameter("PhysicsWorld", "world"), parameter("JobHandle", "inDeps")});
+		lambda.getStatements().add(declaration(declarator("collisionData",creation("NativeList<CollisionData>", argument(access(variable("Allocator"),variable("TempJob")))))));
+		var schedule = function("Schedule", argument(variable("simulation")), refArgument(variable("world")), argument(variable("inDeps")));
+		lambda.getStatements().add(declaration(declarator("job",access(creation("CollisionJob",member("collisionData", variable("collisionData")), member("world", variable("world"))),schedule))));
+		
+		lambda.getStatements().add(statement(access(variable("job"),function("Complete"))));
+
+		var accessA = access(variable("EntityManager"),function("GetBuffer", new String[] {"Collisions"}, argument(access(variable("data"),variable("A")))));
+		var accessB = access(variable("EntityManager"),function("GetBuffer", new String[] {"Collisions"}, argument(access(variable("data"),variable("B")))));
+		var addB = access(accessA, function("Add",argument(creation("Collisions", member("Value", access(variable("data"),variable("B")))))));
+		var addA = access(accessB, function("Add",argument(creation("Collisions", member("Value", access(variable("data"),variable("A")))))));
+		
+		var foreach = foreach("data",variable("collisionData"));
+		var ifA = ifStatement(access(variable("EntityManager"),function("HasComponent", new String[] {"Collisions"}, argument(access(variable("data"),variable("A"))))));
+		ifA.getBranches().get(0).getStatements().add(statement(addB));
+		var ifB = ifStatement(access(variable("EntityManager"),function("HasComponent", new String[] {"Collisions"}, argument(access(variable("data"),variable("B"))))));
+		ifB.getBranches().get(0).getStatements().add(statement(addA));
+		
+		foreach.getStatements().add(ifA);
+		foreach.getStatements().add(ifB);
+		lambda.getStatements().add(foreach);
+		lambda.getStatements().add(statement(access(variable("collisionData"), function("Dispose"))));
+		lambda.getStatements().add(returnStatement(variable("job")));
+		
+		var declareLambda = declaration("SimulationCallbacks.Callback", declarator("testCollisions", lambda));
+		onUpdate.getStatements().add(declareLambda);
+		
+		var jacobians = access(access(variable("SimulationCallbacks"), variable("Phase")),variable("PostSolveJacobians"));
+		var enqueue = access(variable("step"), function("EnqueueCallback", new Argument[] {argument(jacobians), argument(variable("testCollisions")), argument(variable("inputDependencies"))}));
+		onUpdate.getStatements().add(statement(enqueue));
+		
 		onUpdate.getStatements().add(returnStatement(variable("inputDependencies")));
 		
 		clazz.getMembers().add(onCreate);
 		clazz.getMembers().add(onUpdate);
 		clazz.getMembers().add(job);
 		GenericSerializer.generate(unit, csharpModule, fsa, "Unity/Assets/Code/Systems/Engine/DetectCollisions.cs");
-	}
-	
-	CompilationUnit unit ()
-	{
-		return csharp.createCompilationUnit();
-	}
-	
-	CompilationUnit addUsing(CompilationUnit unit, Using using)
-	{
-		unit.getUsings().add(using);
-		return unit;
-	}
-	
-	NamespaceUsing namespaceUsing(String namespace)
-	{
-		var using = csharp.createNamespaceUsing();
-		using.setNamespace(namespace);
-		return using;
-	}
-	
-	StaticUsing staticUsing(String namespace)
-	{
-		var using = csharp.createStaticUsing();
-		using.setNamespace(namespace);
-		return using;
-	}
-	
-	Namespace namespace(String name)
-	{
-		var namespace = csharp.createNamespace();
-		namespace.setName(name);
-		return namespace;
-	}
-	
-	AttributeSection attribute(String name)
-	{
-		var attributeSection = csharp.createAttributeSection();
-		var attribute = csharp.createAttribute();
-		attributeSection.getAttributes().add(attribute);
-		attribute.setName(name);
-		return attributeSection;
-	}
-	
-	AttributeSection attribute(String name, m.modular.Expression[] arguments)
-	{
-		var attributeSection = csharp.createAttributeSection();
-		var attribute = csharp.createAttribute();
-		attributeSection.getAttributes().add(attribute);
-		attribute.setName(name);
-		for (var argument : arguments)
-		{
-			attribute.getPositionalArguments().add(argument);
-		}
-		return attributeSection;
-	}
-	
-	m.csharp.Class clazz(Modifier[] modifiers, String name, String[] superTypes)
-	{
-		var clazz = csharp.createClass();
-		clazz.setName(name);
-		for (var modifier : modifiers)
-		{
-			clazz.getModifiers().add(modifier);
-		}
-		for (var superType : superTypes)
-		{
-			clazz.getSuperTypes().add(superType);
-		}
-		return clazz;
-	}
-	
-	Struct struct(Modifier[] modifiers, String name)
-	{
-		var struct = csharp.createStruct();
-		struct.setName(name);
-		for (var modifier : modifiers)
-		{
-			struct.getModifiers().add(modifier);
-		}
-		return struct;
-	}
-	
-	Struct struct(String name, String[] superTypes)
-	{
-		var struct = csharp.createStruct();
-		struct.setName(name);
-		for (var superType : superTypes)
-		{
-			struct.getSuperTypes().add(superType);
-		}
-		return struct;
-	}
-	
-	Struct struct(String name)
-	{
-		var struct = csharp.createStruct();
-		struct.setName(name);
-		return struct;
-	}
-	
-	Field field(Modifier[] modifiers, String type, Declarator...declarators)
-	{
-		var field = csharp.createField();
-		for (var modifier : modifiers)
-		{
-			field.getModifiers().add(modifier);
-		}
-		field.setType(type);
-		for (var declarator : declarators)
-		{
-			field.getDeclarators().add(declarator);
-		}
-		return field;
-	}
-	
-	Field field(String type, Declarator...declarators)
-	{
-		var field = csharp.createField();
-		field.setType(type);
-		for (var declarator : declarators)
-		{
-			field.getDeclarators().add(declarator);
-		}
-		return field;
-	}
-	
-	Method method(Modifier[] modifiers, String type, String name)
-	{
-		var method = csharp.createMethod();
-		for (var modifier : modifiers)
-		{
-			method.getModifiers().add(modifier);
-		}
-		method.setType(type);
-		method.setName(name);
-		return method;
-	}
-	
-	Method method(Modifier[] modifiers, String type, String name, Parameter[] parameters)
-	{
-		var method = csharp.createMethod();
-		for (var modifier : modifiers)
-		{
-			method.getModifiers().add(modifier);
-		}
-		method.setType(type);
-		method.setName(name);
-		for (var parameter : parameters)
-		{
-			method.getParameters().add(parameter);
-		}
-		return method;
-	}
-	
-	Return returnStatement(m.modular.Expression expression)
-	{
-		var returnStatement = csharp.createReturn();
-		returnStatement.setExpression(expression);
-		return returnStatement;
-	}
-	
-	Default defaultExpression()
-	{
-		return csharp.createDefault();
-	}
-	
-	m.modular.Variable variable(String name)
-	{
-		var variable = modular.createVariable();
-		variable.setName(name);
-		return variable;
-	}
-	
-	ParameterizedFunction function(String name, Argument... arguments)
-	{
-		var function = csharp.createParameterizedFunction();
-		function.setName(name);
-		for (var argument : arguments)
-		{
-			function.getArguments().add(argument);
-		}
-		return function;
-	}
-	
-	ParameterizedFunction function(String name, String[] types, Argument... arguments)
-	{
-		var function = csharp.createParameterizedFunction();
-		function.setName(name);
-		for (var type : types)
-		{
-			function.getTypes().add(type);
-		}
-		for (var argument : arguments)
-		{
-			function.getArguments().add(argument);
-		}
-		return function;
-	}
-	
-	m.modular.LogicalNot not(m.modular.Expression expression)
-	{
-		var not = modular.createLogicalNot();
-		not.setExpression(expression);
-		return not;
-	}
-	
-	Typeof typeof(String type)
-	{
-		var typeof = csharp.createTypeof();
-		typeof.setType(type);
-		return typeof;
-	}
-	
-	m.modular.Brackets brackets(m.modular.Expression expression)
-	{
-		var brackets = modular.createBrackets();
-		brackets.setExpression(expression);
-		return brackets;
-	}
-	
-	AccessExpression access(m.modular.Expression left, m.modular.Expression right)
-	{
-		var access = modular.createAccessExpression();
-		access.setLeft(left);
-		access.setRight(right);
-		return access;
-	}
-	
-	m.modular.Equality equality(m.modular.Expression left, m.modular.Expression right)
-	{
-		var equality = modular.createEquality();
-		equality.setLeft(left);
-		equality.setRight(right);
-		return equality;
-	}
-	
-	m.csharp.Assignment assignment(m.modular.Expression left, m.modular.Expression right)
-	{
-		var assignment = csharp.createAssignment();
-		assignment.setLeft(left);
-		assignment.setRight(right);
-		return assignment;
-	}
-	
-	m.csharp.Assignment assignment(m.modular.Expression left, AssignmentKind kind, m.modular.Expression right)
-	{
-		var assignment = csharp.createAssignment();
-		assignment.setLeft(left);
-		assignment.setRight(right);
-		assignment.setKind(kind);
-		return assignment;
-	}
-	
-	ExpressionStatement statement(m.modular.Expression expression)
-	{
-		var statement = csharp.createExpressionStatement();
-		statement.setExpression(expression);
-		return statement;
-	}
-	
-	Declarator declarator(String name, m.modular.Expression expression)
-	{
-		var declarator = csharp.createDeclarator();
-		declarator.setVariable(name);
-		declarator.setValue(expression);
-		return declarator;
-	}
-	
-	Declarator declarator(String name)
-	{
-		var declarator = csharp.createDeclarator();
-		declarator.setVariable(name);
-		return declarator;
-	}
-	
-	Declaration declaration(String type, Declarator...declarators) 
-	{
-		var declaration = csharp.createDeclaration();
-		declaration.setType(type);
-		for (var declarator : declarators)
-		{
-			declaration.getDeclarators().add(declarator);
-		}
-		return declaration;
-	}
-	
-	Declaration declaration(Declarator...declarators)
-	{
-		var declaration = csharp.createDeclaration();
-		for (var declarator : declarators)
-		{
-			declaration.getDeclarators().add(declarator);
-		}
-		return declaration;
-	}
-	
-	Parameter parameter(String type, String name)
-	{
-		var parameter = csharp.createParameter();
-		parameter.setType(type);
-		parameter.setName(name);
-		return parameter;
-	}
-	
-	Parameter refParameter(String type, String name)
-	{
-		var parameter = csharp.createParameter();
-		parameter.setType(type);
-		parameter.setName(name);
-		parameter.setRef(true);
-		return parameter;
-	}
-	
-	MemberInitializer member(String name, m.modular.Expression value)
-	{
-		var memberInitializer = csharp.createMemberInitializer();
-		memberInitializer.setName(name);
-		memberInitializer.setValue(value);
-		return memberInitializer;
-	}
-	
-	Argument argument(m.modular.Expression expression)
-	{
-		var argument = csharp.createArgument();
-		argument.setValue(expression);
-		return argument;
-	}
-	
-	FloatLiteral floatLiteral(String number)
-	{
-		var floatLiteral = csharp.createFloatLiteral();
-		floatLiteral.setValue(number);
-		return floatLiteral;
-	}
-	
-	m.modular.Comparison comparison(m.modular.Expression left, ComparisonKind kind, m.modular.Expression right)
-	{
-		var comparison = modular.createComparison();
-		comparison.setLeft(left);
-		comparison.setRight(right);
-		comparison.setKind(kind);
-		
-		return comparison;
-	}
-	
-	Creation creation(String name, Argument...arguments)
-	{
-		var creation = csharp.createCreation();
-		creation.setType(name);
-		for (var argument : arguments)
-		{
-			creation.getArguments().add(argument);
-		}
-		return creation;
-	}
-	
-	Creation creation(String name, MemberInitializer...members)
-	{
-		var creation = csharp.createCreation();
-		creation.setType(name);
-		for (var member : members)
-		{
-			creation.getMembers().add(member);
-		}
-		return creation;
 	}
 }
